+++ src/Shared/DC/Scripts/Script.py
@@ -15,16 +15,10 @@
 This provides generic script support
 """
 
-from string import join
-from urllib import quote
-
 from AccessControl.class_init import InitializeClass
-from AccessControl.Permissions import view_management_screens
 from AccessControl.SecurityInfo import ClassSecurityInfo
-from App.special_dtml import DTMLFile
 from DocumentTemplate.DT_Util import TemplateDict
 from OFS.SimpleItem import SimpleItem
-from zExceptions import Redirect
 
 from Shared.DC.Scripts.BindingsUI import BindingsUI
 
@@ -43,19 +37,6 @@ class Script(SimpleItem, BindingsUI):
     func_code = __code__ = None
 
     _Bindings_ns_class = TemplateDict
-
-    security.declareProtected(view_management_screens, 'ZScriptHTML_tryForm')
-    ZScriptHTML_tryForm = DTMLFile('dtml/scriptTry', globals())
-
-    def ZScriptHTML_tryAction(self, REQUEST, argvars):
-        """Apply the test parameters.
-        """
-        vv = []
-        for argvar in argvars:
-            if argvar.value:
-                vv.append("%s=%s" % (quote(argvar.name), quote(argvar.value)))
-        raise Redirect("%s?%s" % (REQUEST['URL1'], join(vv, '&')))
-
     from Signature import _setFuncSignature
 
 InitializeClass(Script)